requirements.txt
appium-python-client==5.3.0
pytest==9.0.2
selenium==4.41.0
pytest-html==4.2.0
